refactor: migrate matrix align to TypeScript
